Fix conversion from ‘double’ to ‘int’

+++ sprint_4/4_allocating_code_to_files/7_templates/carpenter.cpp
@@ -5,7 +5,7 @@
 int Carpenter::CalcShelves(const Wall& wall) const {
     double height = wall.GetHeight();
     double width = wall.GetWidth();
-    return CalcSquare(width, height) / 2;
+    return static_cast<int>(CalcSquare(width, height) / 2);
 }
 
 void Carpenter::InstallDoor(Wall& wall, const Builder& builder) const {
